Fix issue with animation for hidden assets
* create a separate component
* use `scrollIntoView` for button element
